Move category to instances of local plan events, not the types.
Update form handling to bind input fields for date parts to dictionary
and save day, month and year to dictionary rather than table columns.

This allows grouping on inputs from the form to a single event
rather than event per row of inputs. That was making difficult
to group the partical start and end dates.

Added edit form for regulation 18